+++ checkov/terraform/graph_manager.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+import os
 from typing import Type, Any, TYPE_CHECKING, overload
 
 from checkov.common.util.consts import DEFAULT_EXTERNAL_MODULES_DIR
@@ -31,7 +32,7 @@ class TerraformGraphManager(GraphManager[TerraformLocalGraph, "dict[TFDefinition
         excluded_paths: list[str] | None = None,
         external_modules_download_path: str = DEFAULT_EXTERNAL_MODULES_DIR,
         vars_files: list[str] | None = None,
-    ) -> list[tuple[TerraformLocalGraph, list[dict[TFDefinitionKey, dict[str, Any]]], str]]:
+    ) -> tuple[list[tuple[TerraformLocalGraph, list[dict[TFDefinitionKey, dict[str, Any]]], str]], dict[str, str]]:
         logging.info("Parsing HCL files in source dir to multi graph")
         modules_with_definitions = self.parser.parse_multi_graph_hcl_module(
             source_dir=source_dir,
@@ -44,6 +45,7 @@ class TerraformGraphManager(GraphManager[TerraformLocalGraph, "dict[TFDefinition
         )
 
         graphs: list[tuple[TerraformLocalGraph, list[dict[TFDefinitionKey, dict[str, Any]]], str]] = []
+        resource_subgraph_map: dict[str, str] = {}
         for module, tf_definitions in modules_with_definitions:
             logging.info("Building graph from parsed module")
             local_graph = local_graph_class(module)
@@ -51,8 +53,8 @@ class TerraformGraphManager(GraphManager[TerraformLocalGraph, "dict[TFDefinition
             subgraph_abs_path = module.source_dir
             subgraph_path = subgraph_abs_path[subgraph_abs_path.rindex(source_dir) + len(source_dir) + 1:]
             graphs.append((local_graph, tf_definitions, subgraph_path))
-
-        return graphs
+            self.update_resource_subgraph_map(local_graph, subgraph_path, resource_subgraph_map, source_dir)
+        return graphs, resource_subgraph_map
 
     def build_graph_from_source_directory(
         self,
@@ -118,3 +120,11 @@ class TerraformGraphManager(GraphManager[TerraformLocalGraph, "dict[TFDefinition
             graphs.append((source_path, local_graph))
 
         return graphs
+
+    @staticmethod
+    def update_resource_subgraph_map(
+            local_graph: TerraformLocalGraph, subgraph_path: str, resource_subgraph_map: dict[str, str], source_dir: str
+    ) -> None:
+        for v in local_graph.vertices:
+            resource_id = f"/{os.path.relpath(v.path, source_dir)}:{v.id}"
+            resource_subgraph_map[resource_id] = subgraph_path